renames and doc changes
