+++ plugins/inputs/google_pubsub/pubsub.go
@@ -1,4 +1,4 @@
-package pubsub
+package google_pubsub
 
 import (
 	"cloud.google.com/go/pubsub"
@@ -10,7 +10,6 @@ import (
 	"github.com/influxdata/telegraf/plugins/parsers"
 	"golang.org/x/oauth2/google"
 	"google.golang.org/api/option"
-	"log"
 	"sync"
 )
 
@@ -21,7 +20,7 @@ const defaultMaxUndeliveredMessages = 1000
 
 type PubSub struct {
 	CredentialsFile string `toml:"credentials_file"`
-	Project         string `toml:"google_project"`
+	Project         string `toml:"project"`
 	Subscription    string `toml:"subscription"`
 
 	// Subscription ReceiveSettings
@@ -54,7 +53,7 @@ func (ps *PubSub) Description() string {
 }
 
 func (ps *PubSub) SampleConfig() string {
-	return sampleConfig
+	return fmt.Sprintf(sampleConfig, defaultMaxUndeliveredMessages)
 }
 
 // Gather does nothing for this service input.
@@ -95,7 +94,6 @@ func (ps *PubSub) Start(ac telegraf.Accumulator) error {
 		ps.receiveDelivered(cctx)
 	}()
 
-	log.Printf("Started the PubSub service! Listening to subscriptions %+v", ps.Subscription)
 	return nil
 }
 
@@ -231,23 +229,36 @@ func init() {
 }
 
 const sampleConfig = `
-  ## Name of Google Cloud Platform (GCP) Project owning PubSub subscriptions
-  google_project = "my-project"
+  ## Required. Name of Google Cloud Platform (GCP) Project that owns
+  ## the given PubSub subscription.
+  project = "my-project"
 
-  ## Name of PubSub subscriptions
+  ## Required. Name of PubSub subscription to ingest metrics from.
   subscription = "my-subscription"
 
-  ## Optional filepath for GCP credentials JSON file to authorize calls to 
+  ## Required. Data format to consume.
+  ## Each data format has its own unique set of configuration options.
+  ## Read more about them here:
+  ## https://github.com/influxdata/telegraf/blob/master/docs/DATA_FORMATS_INPUT.md
+  data_format = "influx"
+
+  ## Optional. Filepath for GCP credentials JSON file to authorize calls to 
   ## PubSub APIs. If not set explicitly, Telegraf will attempt to use 
   ## Application Default Credentials, which is preferred. 
   # credentials_file = "path/to/my/creds.json"
 
-  ## Maximum byte length of a message to consume. Larger messages are dropped.
-  ## If less than 0 or unspecified, defaults to unlimited length.
-  max_message_len = 1000000
+  ## Optional. If non-empty, this tag is added to each metrics with the
+  ## subscription ID as its value.
+  # subscription_tag = "sub_id"
 
-  ## Maximum messages to read from PubSub that have not been written to an
-  ## output.  For best throughput set based on the number of metrics within
+  ## Optional. Maximum byte length of a message to consume. 
+  ## Larger messages are dropped with an error. If less than 0 or unspecified, 
+  ## treated as no limit.
+  # max_message_len = 1000000
+
+  ## Optional. Maximum messages to read from PubSub that have not been written 
+  ## to an output. Defaults to %d.
+  ## For best throughput set based on the number of metrics within
   ## each message and the size of the output's metric_batch_size.
   ##
   ## For example, if each message contains 10 metrics and the output 
@@ -260,32 +271,27 @@ const sampleConfig = `
   ## Read more about these values:
   ## https://godoc.org/cloud.google.com/go/pubsub#ReceiveSettings
   
-  ## Maximum number of seconds for which a PubSub subscription
+  ## Optional. Maximum number of seconds for which a PubSub subscription
   ## should auto-extend the PubSub ACK deadline for each message. If less than
   ## 0, auto-extension is disabled.
   # max_extension = 0
 
-  ## Maximum number of unprocessed messages in PubSub (unacknowledged but not 
-  ## yet expired in PubSub). A value of 0 is treated as the default 
-  ## PubSub value. Negative values will be treated as unlimited.
+  ## Optional. Maximum number of unprocessed messages in PubSub 
+  ## (unacknowledged but not yet expired in PubSub). 
+  ## A value of 0 is treated as the default PubSub value. 
+  ## Negative values will be treated as unlimited.
   # max_outstanding_messages = 0
 
-  ## Maximum size in bytes of unprocessed messages in PubSub 
+  ## Optional. Maximum size in bytes of unprocessed messages in PubSub 
   ## (unacknowledged but not yet expired in PubSub). 
-  ## A value of 0 is treated as the default PubSub value.
+  ## A value of 0 is treated as the default PubSub value. 
   ## Negative values will be treated as unlimited.
   # max_outstanding_bytes = 0
 
-  ## Max number of goroutines a PubSub Subscription receiver can spawn 
+  ## Optional. Max number of goroutines a PubSub Subscription receiver can spawn 
   ## to pull messages from PubSub concurrently. This limit applies to each 
-  ## subscription separately and is treated as the PubSub default if less than 1.
-  ## Note this setting does not limit the number of messages that can be 
-  ## processed concurrently (use "max_outstanding_messages" instead)
+  ## subscription separately and is treated as the PubSub default if less than 
+  ## 1. Note this setting does not limit the number of messages that can be 
+  ## processed concurrently (use "max_outstanding_messages" instead).
   # max_receiver_go_routines = 0
-
-  ## Data format to consume.
-  ## Each data format has its own unique set of configuration options.
-  ## Read more about them here:
-  ## https://github.com/influxdata/telegraf/blob/master/docs/DATA_FORMATS_INPUT.md
-  data_format = "influx"
 `
